Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -3,7 +3,7 @@ int main(){
     float pi = 3.14;
     int r;
     scanf("%d",&r);
-    float area = pi*r*r;
+    float area = (float)pi*r*r;
     printf("Area: %.2f",area);
     return 0;
 }
